Release: v2025.07.18

+++ root/app/extended-scripts/artists_folders_merger.py
@@ -38,39 +38,41 @@ def process_artists_top_level(base_directory, duplicate_folders_directory, cache
 
         # process each group in the artist map
         for base_name, artist_group in artist_map.items():
-            # create a directory in /config/cache for the base name
-            cache_folder = os.path.join(cache_directory, base_name)
-            os.makedirs(cache_folder, exist_ok=True)
+            # Only process if there is more than one folder for the same artist
+            if len(artist_group) > 1:
+                # create a directory in /config/cache for the base name
+                cache_folder = os.path.join(cache_directory, base_name)
+                os.makedirs(cache_folder, exist_ok=True)
 
-            # copy content of all grouped folders to the cache directory
-            for folder in artist_group:
-                for item in os.listdir(folder):
-                    source_path = os.path.join(folder, item)
-                    target_path = os.path.join(cache_folder, item)
+                # copy content of all grouped folders to the cache directory
+                for folder in artist_group:
+                    for item in os.listdir(folder):
+                        source_path = os.path.join(folder, item)
+                        target_path = os.path.join(cache_folder, item)
 
-                    if os.path.isdir(source_path):
-                        logging.info(f"copying directory: {source_path} -> {target_path}")
-                        shutil.copytree(source_path, target_path, dirs_exist_ok=True)
-                    else:
-                        logging.info(f"copying file: {source_path} -> {target_path}")
-                        shutil.copy2(source_path, target_path)
+                        if os.path.isdir(source_path):
+                            logging.info(f"copying directory: {source_path} -> {target_path}")
+                            shutil.copytree(source_path, target_path, dirs_exist_ok=True)
+                        else:
+                            logging.info(f"copying file: {source_path} -> {target_path}")
+                            shutil.copy2(source_path, target_path)
 
-            # move original folders to the duplicate folder directory
-            for folder in artist_group:
-                target_folder = os.path.join(duplicate_folders_directory, os.path.basename(folder))
+                # move original folders to the duplicate folder directory
+                for folder in artist_group:
+                    target_folder = os.path.join(duplicate_folders_directory, os.path.basename(folder))
 
-                # verificar se o diretório de destino já existe
-                if os.path.exists(target_folder):
-                    logging.info(f"destination path '{target_folder}' already exists. removing and overwriting.")
-                    shutil.rmtree(target_folder)  # Remove o diretório existente
+                    # verificar se o diretório de destino já existe
+                    if os.path.exists(target_folder):
+                        logging.info(f"destination path '{target_folder}' already exists. removing and overwriting.")
+                        shutil.rmtree(target_folder)  # Remove o diretório existente
 
-                logging.info(f"moving original folder: {folder} -> {target_folder}")
-                shutil.move(folder, target_folder)
+                    logging.info(f"moving original folder: {folder} -> {target_folder}")
+                    shutil.move(folder, target_folder)
 
-            # move the merged cache folder back to the original directory
-            target_path = os.path.join(base_directory, base_name)
-            logging.info(f"moving merged directory: {cache_folder} -> {target_path}")
-            shutil.move(cache_folder, target_path)
+                # move the merged cache folder back to the original directory
+                target_path = os.path.join(base_directory, base_name)
+                logging.info(f"moving merged directory: {cache_folder} -> {target_path}")
+                shutil.move(cache_folder, target_path)
 
         logging.info("artist directory processing completed successfully.")
     except Exception as error:
